+++ GetThePicture/Copybook/Provider/DataProvider.cs
@@ -1,3 +1,4 @@
+using System.Text;
 using GetThePicture.Copybook.Compiler;
 using GetThePicture.Copybook.Compiler.Layout;
 using GetThePicture.Copybook.Resolver;
@@ -10,8 +11,8 @@ namespace GetThePicture.Copybook.Provider;
 /// </summary>
 public sealed class DataProvider : IDataProvider
 {
-    private readonly Lazy<CbLayout> _layout;
-    private readonly Lazy<CbStorage> _storage;
+    private readonly CbLayout _layout;
+    private readonly CbStorage _storage;
 
     /// <summary>
     /// Initializes a new instance of <see cref="DataProvider"/> using a <see cref="StreamReader"/> that reads a COBOL Copybook.
@@ -19,8 +20,16 @@ public sealed class DataProvider : IDataProvider
     /// <param name="reader">The <see cref="StreamReader"/> for the Copybook file (e.g., .cpy or .cbl).</param>
     public DataProvider(StreamReader reader)
     {
-        _layout  = new Lazy<CbLayout> (() => CbCompiler.FromStreamReader(reader));
-        _storage = new Lazy<CbStorage>(() => CbResolver.FromLayout(_layout.Value));
+        _layout  = CbCompiler.FromStreamReader(reader);
+        _storage = CbResolver.FromLayout(_layout);
+    }
+
+    public DataProvider(string fileName, Encoding encoding)
+    {
+        using var reader = new StreamReader(fileName, encoding);
+
+        _layout  = CbCompiler.FromStreamReader(reader);
+        _storage = CbResolver.FromLayout(_layout);
     }
 
     /// <summary>
@@ -28,12 +37,12 @@ public sealed class DataProvider : IDataProvider
     /// This object is lazily initialized on first access.
     /// </summary>
     /// <returns>The <see cref="CbLayout"/> instance.</returns>
-    public CbLayout GetLayout() => _layout.Value;
+    public CbLayout GetLayout() => _layout;
 
     /// <summary>
     /// Gets the <see cref="CbStorage"/> object representing the resolved storage map of the Copybook.
     /// This object depends on <see cref="GetLayout"/> and is lazily initialized on first access.
     /// </summary>
     /// <returns>The <see cref="CbStorage"/> instance.</returns>
-    public CbStorage GetStorage() => _storage.Value;
+    public CbStorage GetStorage() => _storage;
 }